Added comments for game infra blocking
Added in comments for the game infrastructure. These sites act as a content delivery network (CDN) for many popular game websites. Even if we can't find websites which have games on them, they are likely to rely on one of these to show videos, pages, and serve javascript. If they are not able to connect to the CDN, the games will not load